requirements/production.txt
Django==4.1.1
django-cors-headers==3.13.0
django-filter==22.1        
djangorestframework==3.14.0
drf-spectacular==0.24.2
Pillow==9.4.0
python-decouple==3.6
requests==2.28.2
